Разбил код по модулям и Убрал магические числа
