+++ .gitignore
@@ -10,5 +10,7 @@ Thumbs.db
 # Вспомогательные файлы
 *.log*
 node_modules/
+public/bundle.js
+public/bundle.js.map
 
 # Файлы сборки (допишите самостоятельно)
