+++ src/components/footer-movies-amount.js
@@ -1,0 +1,7 @@
+export const getFooterMoviesAmount = () => {
+  return (
+    `<section class="footer__statistics">
+      <p>0 movies inside</p>
+    </section>`
+  );
+};
