+++ src/components/load-more-button.js
@@ -1,0 +1,5 @@
+export const getShowMoreFilmsButton = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
