+++ src/components/user-rank.js
@@ -1,0 +1,8 @@
+export const getUserRank = () => {
+  return (
+    `<section class="header__profile profile">
+      <p class="profile__rating">Sci-Fighter</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>`
+  );
+};
